requirements.txt
Django==1.4
South==0.7.5
lxml==2.3.4
requests==0.13.3
celery-with-redis
django-celery==2.5.5
